move summary collection to one file; update fullTest comments
