+++ utils/summary/collectSessions.js
@@ -1,14 +1,14 @@
-const utils = require('../../utils')
+const type = require('../type')
 
 const collectSessions = (fileList, options) => {
   let sessions = []
   Object.keys(fileList).forEach(key => {
     const file = fileList[key]
     if (
-      !utils.type.file.isStimuliData(file.relativePath) &&
-      utils.type.isBIDS(file.relativePath, options.bep006, options.bep010)
+      !type.file.isStimuliData(file.relativePath) &&
+      type.isBIDS(file.relativePath, options.bep006, options.bep010)
     ) {
-      const pathValues = utils.type.getPathValues(file.relativePath)
+      const pathValues = type.getPathValues(file.relativePath)
       const isEmptyRoom = pathValues.sub && pathValues.sub == 'emptyroom'
 
       if (
