+++ utils/summary/collectSummary.js
@@ -1,0 +1,38 @@
+const files = require('../files')
+const collectModalities = require('./collectModalities')
+const collectSessions = require('./collectSessions')
+const collectSubjects = require('./collectSubjects')
+
+const collectSummary = (fileList, options) => {
+  let summary = {
+    sessions: [],
+    subjects: [],
+    tasks: [],
+    modalities: [],
+    totalFiles: Object.keys(fileList).length,
+    size: 0,
+  }
+
+  //collect file directory statistics
+  summary.size = files.collectDirectoryStatistics(fileList)
+
+  // remove ignored files from list:
+  Object.keys(fileList).forEach(function(key) {
+    if (fileList[key].ignore) {
+      delete fileList[key]
+    }
+  })
+
+  // collect modalities for summary
+  summary.modalities = collectModalities(fileList)
+
+  // collect subjects
+  summary.subjects = collectSubjects(fileList, options)
+
+  // collect sessions
+  summary.sessions = collectSessions(fileList, options)
+
+  return summary
+}
+
+module.exports = collectSummary
