+++ utils/summary/index.js
@@ -1,0 +1,11 @@
+const collectSummary = require('./collectSummary')
+const collectModalities = require('./collectModalities')
+const collectSessions = require('./collectSessions')
+const collectSubjects = require('./collectSubjects')
+
+module.exports = {
+  collectSummary,
+  collectModalities,
+  collectSessions,
+  collectSubjects,
+}
