+++ validators/bids/fullTest.js
@@ -12,8 +12,6 @@ const checkAnyDataPresent = require('../checkAnyDataPresent')
 const headerFields = require('../headerFields')
 const subSesMismatchTest = require('./subSesMismatchTest')
 const groupFileTypes = require('./groupFileTypes')
-const collectModalities = require('./collectModalities')
-const collectSessions = require('./collectSessions')
 const subjects = require('./subjects')
 const checkDatasetDescription = require('./checkDatasetDescription')
 
@@ -42,17 +40,7 @@ const fullTest = (fileList, options, callback) => {
 
   let tsvs = []
 
-  let summary = {
-    sessions: [],
-    subjects: [],
-    tasks: [],
-    modalities: [],
-    totalFiles: Object.keys(fileList).length,
-    size: 0,
-  }
-
-  //collect file directory statistics
-  summary.size = utils.files.collectDirectoryStatistics(fileList)
+  let summary = utils.collectSummary(fileList, self.options)
 
   // remove ignored files from list:
   Object.keys(fileList).forEach(function(key) {
@@ -80,9 +68,6 @@ const fullTest = (fileList, options, callback) => {
     )
   })
 
-  // collect modalities for summary
-  summary.modalities = collectModalities(fileList)
-
   // check if dataset contains T1w
   if (summary.modalities.indexOf('T1w') < 0) {
     self.issues.push(
@@ -92,8 +77,7 @@ const fullTest = (fileList, options, callback) => {
     )
   }
 
-  // SECTION: TSV
-  // collect tsv issues
+  // TSV validation
   tsv
     .validate(
       files.tsv,
@@ -107,48 +91,35 @@ const fullTest = (fileList, options, callback) => {
     .then(tsvIssues => {
       self.issues = self.issues.concat(tsvIssues)
 
-      // SECTION: BVEC
-      // collect bvec issues
+      // Bvec validation
       bvec.validate(files.bvec, bContentsDict).then(bvecIssues => {
         self.issues = self.issues.concat(bvecIssues)
 
-        // SECTION: BVAL
-        // collect bval issues
+        // Bval validation
         bval.validate(files.bval, bContentsDict).then(bvalIssues => {
           self.issues = self.issues.concat(bvalIssues)
 
-          // SECTION: JSON
-          // load json files and construct a contents object with field, value pairs
+          // Load json files and construct a contents object with field, value pairs
           json
             .load(files.json, jsonFiles, jsonContentsDict)
             .then(jsonLoadIssues => {
               self.issues = self.issues.concat(jsonLoadIssues)
 
-              // check for at least one subject
+              // Check for at least one subject
               const noSubjectIssues = subjects.atLeastOneSubject(fileList)
               self.issues = self.issues.concat(noSubjectIssues)
 
-              // check for datasetDescription file in the proper place
+              // Check for datasetDescription file in the proper place
               const datasetDescriptionIssues = checkDatasetDescription(fileList)
               self.issues = self.issues.concat(datasetDescriptionIssues)
 
-              // collect subjects
-              summary.subjects = subjects.collectSubjects(
-                fileList,
-                self.options,
-              )
-
-              // collect sessions
-              summary.sessions = collectSessions(fileList, self.options)
-
-              // validate json files and contents
+              // Validate json files and contents
               json
                 .validate(jsonFiles, fileList, jsonContentsDict, summary)
                 .then(jsonIssues => {
                   self.issues = self.issues.concat(jsonIssues)
 
-                  // SECTION: NIFTI
-                  // general nifti validation
+                  // Nifti validation
                   NIFTI.validate(
                     files.nifti,
                     fileList,
@@ -160,8 +131,7 @@ const fullTest = (fileList, options, callback) => {
                   ).then(niftiIssues => {
                     self.issues = self.issues.concat(niftiIssues)
 
-                    // SECTION: PARTICIPANTS
-                    // issues related to participants not listed in the subjects list
+                    // Issues related to participants not listed in the subjects list
                     const participantsInSubjectsIssues = subjects.participantsInSubjects(
                       participants,
                       summary.subjects,
@@ -170,17 +140,16 @@ const fullTest = (fileList, options, callback) => {
                       participantsInSubjectsIssues,
                     )
 
-                    //check for equal number of participants from ./phenotype/*.tsv and participants in dataset
+                    // Check for equal number of participants from ./phenotype/*.tsv and participants in dataset
                     const phenotypeIssues = tsv.checkPhenotype(
                       phenotypeParticipants,
                       summary,
                     )
                     self.issues = self.issues.concat(phenotypeIssues)
 
-                    // validate nii header fields
+                    // Validate nii header fields
                     self.issues = self.issues.concat(headerFields(headers))
 
-                    // SECTION: EVENTS
                     // Events validation
                     stimuli.directory = files.stimuli
                     const eventsIssues = Events.validateEvents(
@@ -192,23 +161,25 @@ const fullTest = (fileList, options, callback) => {
                     )
                     self.issues = self.issues.concat(eventsIssues)
 
-                    // SECTION: TSV COLUMNS
-                    // validate custom fields in all TSVs and add any issues to the list
+                    // Validate custom fields in all TSVs and add any issues to the list
                     self.issues = self.issues.concat(
                       tsv.validateTsvColumns(tsvs, jsonContentsDict),
                     )
 
-                    // SECTION: SESSIONS
-                    // validation session files
+                    // Validate session files
                     self.issues = self.issues.concat(session(fileList))
 
-                    // SECTION: SUBJECT DATA PRESENT
+                    // Determine if each subject has data present
                     self.issues = self.issues.concat(
                       checkAnyDataPresent(fileList, summary.subjects),
                     )
+
+                    // Group summary modalities
                     summary.modalities = utils.modalities.group(
                       summary.modalities,
                     )
+
+                    // Format issues
                     const issues = utils.issues.format(
                       self.issues,
                       summary,
