+++ validators/bids/subjects.js
@@ -38,32 +38,7 @@ const atLeastOneSubject = fileList => {
   return issues
 }
 
-const collectSubjects = (fileList, options) => {
-  let subjects = []
-  const fileKeys = Object.keys(fileList)
-  fileKeys.forEach(key => {
-    const file = fileList[key]
-    if (
-      !utils.type.file.isStimuliData(file.relativePath) &&
-      utils.type.isBIDS(file.relativePath, options.bep006, options.bep010)
-    ) {
-      const pathValues = utils.type.getPathValues(file.relativePath)
-      const isEmptyRoom = pathValues.sub && pathValues.sub == 'emptyroom'
-
-      if (
-        pathValues.sub &&
-        subjects.indexOf(pathValues.sub) === -1 &&
-        !isEmptyRoom
-      ) {
-        subjects.push(pathValues.sub)
-      }
-    }
-  })
-  return subjects
-}
-
 module.exports = {
   participantsInSubjects,
   atLeastOneSubject,
-  collectSubjects,
 }
